network: Indicate the maximum password length accepted for wifi hotspots
The Wifi Hotspot setup dialog will error "Must have a minimum of 8 characters"
for all the situations in which nm_utils_wpa_psk_valid and nm_utils_wep_key_valid
report a given password to be invalid.

The case when this error is most obvious is when you enter a very lengthy password
and the dialog shows the very same error about the minimum password length.

These changes make the dialog account for the Maximum password length as well. It
assumes 16 characters for WEP and 63 characters for WPA/WPA2.